// onecgiar-pr-server/src/shared/constants/result-type.enum.ts
export enum ResultTypeEnum {
  POLICY_CHANGE = 1,
  INNOVATION_USE = 2,
  CAPACITY_CHANGE = 3,
  OTHER_OUTCOME = 4,
  CAPACITY_SHARING_FOR_DEVELOPMENT = 5,
  KNOWLEDGE_PRODUCT = 6,
  INNOVATION_DEVELOPMENT = 7,
  OTHER_OUTPUT = 8,
  IMPACT_CONTRIBUTION = 9,
  INNOVATION_USE_IPSR = 10,
  COMPLEMENTARY_INNOVATION = 11,
}